# docs/specs/afal-http-openapi-draft.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AFAL Phase 1 HTTP Capability Contract",
    "version": "0.1.0-draft",
    "summary": "OpenAPI-style draft for the framework-free AFAL capability transport layer",
    "description": "This draft describes the Phase 1 HTTP surface above the AFAL API adapter.\n\nIt is intentionally narrow and tracks the current implementation under `backend/afal/http/`.\nThe goal is to freeze route names, request envelopes, status-code behavior, and top-level\nresponse semantics before choosing an HTTP server framework.\n\nPhase 1 scope in this draft:\n- request payment approval capability\n- execute payment capability\n- request resource approval capability\n- settle resource usage capability\n- approval session lookup, callback result application, authorization resume, and AFAL action resume\n- stable requestRef handling\n- structured success/error envelopes\n\nCurrent behavior note:\n- `request-payment-approval` and `request-resource-approval` expose top-level `pending-approval` AFAL responses\n- `resume-approved-action` resumes a persisted approved action into settlement and receipt creation\n- `execute-payment` and `settle-resource-usage` still represent the synchronous seeded flow path\n"
  },
  "servers": [
    {
      "url": "https://api.afal.local",
      "description": "Placeholder server for contract drafting only"
    }
  ],
...
  "paths": {
    "/capabilities/request-payment-approval": {
      "post": {
        "tags": [
          "Capabilities"
        ],
        "operationId": "requestPaymentApproval",
        "summary": "Create a top-level pending payment approval request",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestPaymentApprovalHttpRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payment request entered pending approval state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestPaymentApprovalSuccessResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestResponse"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalErrorResponse"
          }
        }
      }
    },
    "/capabilities/execute-payment": {
      "post": {
        "tags": [
          "Capabilities"
        ],
        "operationId": "executePayment",
        "summary": "Execute a payment capability request",
        "description": "Accepts a structured payment capability request and returns either a fully structured\npayment flow result or a structured AFAL error envelope.\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecutePaymentHttpRequest"
              },
              "examples": {
                "canonical-payment": {
                  "summary": "Canonical Phase 1 payment request",
                  "value": {
                    "requestRef": "req-0001",
                    "input": {
                      "requestRef": "req-0001",
                      "monetaryBudgetRef": "budg-money-001",
                      "intent": {
                        "intentId": "payint-0001",
                        "schemaVersion": "0.1",
                        "intentType": "payment",
                        "payer": {
                          "agentDid": "did:afal:agent:payment-agent-01",
                          "accountId": "acct-agent-001"
                        },
                        "payee": {
                          "payeeDid": "did:afal:agent:fraud-service-01",
                          "settlementAddress": "0xFRAUDSERVICEPAYEE"
                        },
                        "asset": "USDC",
                        "amount": "45.00",
                        "chain": "base",
                        "purpose": {
                          "category": "service-payment",
                          "description": "fraud detection request",
                          "referenceId": "svc-req-abc123"
                        },
                        "mandateRef": "mnd-0001",
                        "policyRef": "cred-policy-0001",
                        "executionMode": "pre-authorized",
                        "challengeState": "required",
                        "status": "created",
                        "expiresAt": "2026-03-24T12:10:00Z",
                        "nonce": "n-0001",
                        "createdAt": "2026-03-24T12:05:00Z"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payment flow completed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutePaymentSuccessResponse"
                },
                "examples": {
                  "canonical-payment-success": {
                    "summary": "Canonical payment success response",
                    "value": {
                      "ok": true,
                      "capability": "executePayment",
                      "requestRef": "req-0001",
                      "statusCode": 200,
                      "data": {
                        "intent": {
                          "intentId": "payint-0001",
                          "status": "settled",
                          "challengeState": "approved",
                          "settlementRef": "stl-0001",
                          "receiptRef": "rcpt-pay-0001"
                        },
                        "initialDecision": {
                          "decisionId": "dec-0001",
                          "result": "challenge-required"
                        },
                        "finalDecision": {
                          "decisionId": "dec-0002",
                          "result": "approved"
                        },
                        "settlement": {
                          "settlementId": "stl-0001",
                          "status": "settled"
                        },
                        "paymentReceipt": {
                          "receiptId": "rcpt-pay-0001",
                          "receiptType": "payment",
                          "status": "final"
                        },
                        "capabilityResponse": {
                          "responseId": "cap-0001",
                          "result": "approved"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestResponse"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "409": {
            "$ref": "#/components/responses/ConflictResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalErrorResponse"
          }
        }
      }
    },
    "/capabilities/request-resource-approval": {
      "post": {
        "tags": [
          "Capabilities"
        ],
        "operationId": "requestResourceApproval",
        "summary": "Create a top-level pending resource approval request",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestResourceApprovalHttpRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resource request entered pending approval state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestResourceApprovalSuccessResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestResponse"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalErrorResponse"
          }
        }
      }
    },
    "/capabilities/settle-resource-usage": {
      "post": {
        "tags": [
          "Capabilities"
        ],
        "operationId": "settleResourceUsage",
        "summary": "Execute a resource settlement capability request",
        "description": "Accepts a structured resource capability request and returns either a fully structured\nresource flow result or a structured AFAL error envelope.\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettleResourceUsageHttpRequest"
              },
              "examples": {
                "canonical-resource": {
                  "summary": "Canonical Phase 1 resource settlement request",
                  "value": {
                    "requestRef": "req-1001",
                    "input": {
                      "requestRef": "req-1001",
                      "resourceBudgetRef": "budg-res-001",
                      "resourceQuotaRef": "quota-001",
                      "intent": {
                        "intentId": "resint-0001",
                        "schemaVersion": "0.1",
                        "intentType": "resource",
                        "requester": {
                          "agentDid": "did:afal:agent:research-agent-01",
                          "accountId": "acct-agent-002"
                        },
                        "provider": {
                          "providerId": "provider-openai",
                          "providerDid": "did:afal:institution:provider-openai"
                        },
                        "resource": {
                          "resourceClass": "inference",
                          "resourceUnit": "tokens",
                          "quantity": 500000
                        },
                        "pricing": {
                          "maxSpend": "18.50",
                          "asset": "USDC"
                        },
                        "budgetSource": {
                          "type": "ats-budget",
                          "reference": "budg-res-001"
                        },
                        "mandateRef": "mnd-0002",
                        "policyRef": "cred-policy-0002",
                        "executionMode": "pre-authorized",
                        "challengeState": "required",
                        "status": "created",
                        "expiresAt": "2026-03-24T12:30:00Z",
                        "nonce": "n-2001",
                        "createdAt": "2026-03-24T12:20:00Z"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resource settlement flow completed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettleResourceUsageSuccessResponse"
                },
                "examples": {
                  "canonical-resource-success": {
                    "summary": "Canonical resource settlement success response",
                    "value": {
                      "ok": true,
                      "capability": "settleResourceUsage",
                      "requestRef": "req-1001",
                      "statusCode": 200,
                      "data": {
                        "intent": {
                          "intentId": "resint-0001",
                          "status": "settled",
                          "challengeState": "approved",
                          "settlementRef": "stl-1001",
                          "usageReceiptRef": "usage-1001"
                        },
                        "initialDecision": {
                          "decisionId": "dec-1001",
                          "result": "challenge-required"
                        },
                        "finalDecision": {
                          "decisionId": "dec-1002",
                          "result": "approved"
                        },
                        "usageConfirmation": {
                          "usageReceiptRef": "usage-1001"
                        },
                        "settlement": {
                          "settlementId": "stl-1001",
                          "status": "settled"
                        },
                        "resourceReceipt": {
                          "receiptId": "rcpt-res-1001",
                          "receiptType": "resource",
                          "status": "final"
                        },
                        "capabilityResponse": {
                          "responseId": "cap-1001",
                          "result": "approved"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestResponse"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "409": {
            "$ref": "#/components/responses/ConflictResponse"
          },
          "502": {
            "$ref": "#/components/responses/ProviderFailureResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalErrorResponse"
          }
        }
      }
    },
    "/approval-sessions/get": {
      "post": {
        "tags": [
          "Approvals"
        ],
        "operationId": "getApprovalSession",
        "summary": "Read a persisted approval session",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetApprovalSessionHttpRequest"
              },
              "examples": {
                "canonical-get-approval-session": {
                  "value": {
                    "requestRef": "req-approval-session-001",
                    "input": {
                      "approvalSessionRef": "aps-chall-0001"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Approval session resolved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetApprovalSessionSuccessResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalErrorResponse"
          }
        }
      }
    },
    "/approval-sessions/apply-result": {
      "post": {
        "tags": [
          "Approvals"
        ],
        "operationId": "applyApprovalResult",
        "summary": "Apply a trusted-surface approval callback result",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyApprovalResultHttpRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Approval result was persisted against the approval session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplyApprovalResultSuccessResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalErrorResponse"
          }
        }
      }
    },
    "/approval-sessions/resume": {
      "post": {
        "tags": [
          "Approvals"
        ],
        "operationId": "resumeApprovalSession",
        "summary": "Resume authorization after an approval result is present",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeApprovalSessionHttpRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Approval session resumed and final authorization was produced",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeApprovalSessionSuccessResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalErrorResponse"
          }
        }
      }
    },
    "/approval-sessions/resume-action": {
      "post": {
        "tags": [
          "Approvals"
        ],
        "operationId": "resumeApprovedAction",
        "summary": "Resume an approved AFAL action into settlement, receipt, and final response creation",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeApprovedActionHttpRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Approved action resumed to final AFAL settlement output",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeApprovedActionSuccessResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestResponse"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "409": {
            "$ref": "#/components/responses/ConflictResponse"
          },
          "502": {
            "$ref": "#/components/responses/ProviderFailureResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalErrorResponse"
          }
        }
      }
    }
  },
  "components": {
...
    "schemas": {
      "RequestPaymentApprovalHttpRequest": {
        "$ref": "#/components/schemas/ExecutePaymentHttpRequest"
      },
...
      "RequestResourceApprovalHttpRequest": {
        "$ref": "#/components/schemas/SettleResourceUsageHttpRequest"
      },
...
      "ResumeApprovedActionHttpRequest": {
        "$ref": "#/components/schemas/ResumeApprovalSessionHttpRequest"
      },
...
      "RequestPaymentApprovalSuccessResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "ok",
          "capability",
          "requestRef",
          "statusCode",
          "data"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true
          },
          "capability": {
            "type": "string",
            "const": "requestPaymentApproval"
          },
          "requestRef": {
            "type": "string"
          },
          "statusCode": {
            "type": "integer",
            "const": 200
          },
          "data": {
            "$ref": "#/components/schemas/PaymentApprovalRequestOutput"
          }
        }
      },
...
    }
  }
}